feat: Improve user profile setup flow and home page data loading

Add comprehensive user profile setup with multi-step form including:

Basic info (name, email, language)
Personal details (DOB, gender, height/weight with metric/imperial toggle)
Training preferences (goal selection with auto goal event date)
Health information (injury notes with persistent text controllers)
Fix text input focus issues in profile setup forms by using persistent TextEditingController instances

Update navigation to use GoRouter consistently throughout the app

Replace Navigator.pushReplacementNamed with context.go()
Fix route conflicts between declarative and imperative navigation
Enhance AppButton component with proper state management:

Add disabled state handling with custom colors
Implement loading state with spinner
Consistent sizing and styling across app
Improve home page user data loading:

Add proper null safety handling for user profile checks
Distinguish between "no profile" and "loading error" states
Remove artificial loading delays in favor of proper async handling
Fix DateTime handling issues (remove incorrect .toDate() calls)
Update color scheme for dark theme compatibility:

Fix text colors on dark backgrounds
Improve visibility of form elements and buttons
Ensure proper contrast ratios
Remove Quick Run button, keep only scheduled training runs

Add skip functionality in profile setup (skip to last step, not shown on basic info)

Fix validation and error handling throughout profile setup flow

Improve code organization:

Separate concerns between UI and business logic
Add proper error messages and user feedback
Implement consistent button styling and behavior
Closes #15

Breaking changes:

User profile setup now required before accessing main app features
Navigation routes updated to use GoRouter exclusively

+++ lib/core/services/run_service.dart
@@ -140,7 +140,7 @@ class RunService {
       final runs = await getRunsInDateRange(userId, startDate, endDate);
 
       int totalRuns = runs.length;
-      int totalDistance = 0;
+      double totalDistance = 0.0;
       int totalDuration = 0;
 
       for (var run in runs) {
